Interface: updated one_row, all_col, do, cdo interface
Heuristic: if first arg is not ref (so I guess it's SQL-statement), and second arg is ref
(arrayref with sprintf args), then use sprintf to fill SQL-statement with values from
second args, the rest of args are bind params.

So you can write
    $self->do("SELECT * FROM %s WHERE id = ?", ['table'], 15);
or
    $self->do("SELECT * FROM table WHERE id = ?", 15);
and it will work transparently in both cases.